Move WareHouse to core submodule
Update WareHourse imports as well
