+++ apps/node/src/app/main/model_centric/syft_assets/protocol_manager.py
@@ -2,7 +2,7 @@
 from ...core.exceptions import ProtocolNotFoundError
 
 # PyGrid imports
-from ...storage.warehouse import Warehouse
+from ...core.warehouse import Warehouse
 from .protocol import Protocol
