+++ apps/node/src/app/main/model_centric/workers/worker_manager.py
@@ -4,7 +4,7 @@ import logging
 from ...core.exceptions import WorkerNotFoundError
 
 # PyGrid imports
-from ...storage.warehouse import Warehouse
+from ...core.warehouse import Warehouse
 from .worker import Worker
